+++ R/make_roster_small.R
@@ -20,7 +20,7 @@
 #'
 #' @examples
 #' # Load a sample roster from the package's extdata directory
-#' roster_file <- system.file("extdata/roster.csv", package = "zoomstudentengagement")
+#' roster_file <- system.file("extdata/roster.csv", package = "engager")
 #' roster_df <- readr::read_csv(roster_file, show_col_types = FALSE)
 #' make_roster_small(roster_df = roster_df)
 make_roster_small <- function(roster_df = NULL) {